Revert "js: Workaround for function with custom prototype"
This reverts commit 72c0298c5f9df9036ed67fd504db84cbc028daaa.
The warning that was the reason for this code in the first place was
removed in SpiderMonkey 52. According to
https://bugzilla.mozilla.org/show_bug.cgi?id=1049041 the warning was out
of date, and there are only performance effects "if the object escapes to
somewhere interesting before getting its final __proto__".

In addition, client code should move to ES6 classes which do not have
this performance problem.

Edited to resolve conflicts due to code being moved around in the
meantime, and to use Object.setPrototypeOf() instead of the deprecated
__proto__ property.