class: Stop using custom __name__ property
Since ES6 classes are functions, if you declare 'class Foo {}' then
'Foo.name === "Foo"'. We had a custom __name__ property on Lang.Class
instances, but instead mimic the normal JS class and put a name property
on the class.

That is, the equivalent to 'instance.__name__' is now
'instance.constructor.name'.

This could potentially break client code if any was using this, but the
double underscore made it clear enough that the property should have been
internal.